backend/requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
joblib

numpy
pandas
scikit-learn
scipy
threadpoolctl
